+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Function to retrieve game name from game enum, e.g., `G1830` -> `1830`.
+- Functions to retrieve transcript identifier and ID from filename.
 
 ### Changed